+++ src/Symfony/Component/Console/Command/InvokableCommand.php
@@ -15,6 +15,7 @@ use Symfony\Component\Console\Application;
 use Symfony\Component\Console\ArgumentResolver\ArgumentResolver;
 use Symfony\Component\Console\ArgumentResolver\ArgumentResolverInterface;
 use Symfony\Component\Console\Attribute\Argument;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Attribute\Interact;
 use Symfony\Component\Console\Attribute\MapInput;
 use Symfony\Component\Console\Attribute\Option;
@@ -105,6 +106,19 @@ class InvokableCommand implements SignalableCommandInterface
                     $definition->addOption($option->toInputOption());
                 }
             }
+        }
+
+        // the options listed in the attribute come after the ones the parameters declare
+        $class = $this->invokable->getClosureScopeClass();
+        $name = $this->invokable->getName();
+        $attribute = $class?->hasMethod($name) ? ($class->getMethod($name)->getAttributes(AsCommand::class)[0] ?? null)?->newInstance() : null;
+
+        if (!$attribute && '__invoke' === $name) {
+            $attribute = ($class?->getAttributes(AsCommand::class)[0] ?? null)?->newInstance();
+        }
+
+        foreach ($attribute->options ?? [] as $option) {
+            $definition->addOption($option);
         }
     }